+++ plataforma_web/usuarios/schemas.py
@@ -2,6 +2,7 @@
 Usuarios, esquemas de pydantic
 """
 from pydantic import BaseModel
+from typing import Optional
 
 
 class Usuario(BaseModel):
@@ -17,4 +18,23 @@ class Usuario(BaseModel):
     email: str
     nombres: str
     apellido_paterno: str
-    apellido_materno: str
+    apellido_materno: Optional[str] = None
+
+
+class UsuarioEnBD(Usuario):
+    """Usuario en BD"""
+
+    hashed_password: str
+
+
+class Token(BaseModel):
+    """Token"""
+
+    access_token: str
+    token_type: str
+
+
+class TokenData(BaseModel):
+    """Token data"""
+
+    username: Optional[str] = None
